[trace-expansion] Fix type error tests

+++ firebase-perf/src/jsMain/kotlin/dev/gitlive/firebase/perf/metrics/Trace.kt
@@ -11,9 +11,9 @@ actual class Trace internal constructor(private val js: firebase.performance.Tra
     actual fun stop() = js.stop()
     actual fun getLongMetric(metricName: String) = js.getMetric(metricName).toLong()
 
-    actual fun incrementMetric(metricName: String, incrementBy: Long) = js.incrementMetric(metricName, incrementBy.toInt())
+    actual fun incrementMetric(metricName: String, incrementBy: Long) = js.incrementMetric(metricName, incrementBy)
 
-    actual fun putMetric(metricName: String, value: Long) = js.putMetric(metricName, value.toInt())
+    actual fun putMetric(metricName: String, value: Long) = js.putMetric(metricName, value)
 
     actual fun getAttributes(): Map<String, String> = js.getAttributes()
